separate thought transform logic into a util
use thoughtTransform for hashContext

use hashed key comparison for filtering children
