+++ src/util/hashThought.ts
@@ -1,33 +1,8 @@
 import * as murmurHash3 from 'murmurhash3js'
-import emojiStrip from 'emoji-strip'
-import * as pluralize from 'pluralize'
 import _ from 'lodash'
 import globals from '../globals'
-import { REGEXP_TAGS } from '../constants'
 import { ThoughtHash } from '../types'
-
-/** Converts a string to lowecase. */
-const lower = (s: string) => s.toLowerCase()
-
-/** Trims a string. */
-const trim = (s: string) => s.replace(
-  s.length > 0 && s.replace(/\W/g, '').length > 0 ? /\W/g : /s/g,
-  ''
-)
-
-/** Strips emoji from text. Preserves emoji on its own. */
-const stripEmojiWithText = (s: string) => {
-  const stripped = emojiStrip(s)
-  return stripped.length > 0 ? stripped : s
-}
-
-/** Strips all html tags. */
-const stripTags = (s: string) => s.replace(REGEXP_TAGS, '')
-
-/**
- * Making character 's' will just become an empty value ''.
- * Skip it else it will cause "s" character to have same no of context as empty thoughts in the entire tree. */
-const singularize = (s: string) => s !== 's' ? pluralize.singular(s) : s
+import { thoughtTransform } from './thoughtTransform'
 
 // eslint-disable-next-line jsdoc/require-description-complete-sentence
 /**
@@ -44,12 +19,7 @@ const singularize = (s: string) => s !== 's' ? pluralize.singular(s) : s
  */
 export const hashThought: (s: string) => ThoughtHash = _.memoize(
   (value: string) => _.flow([
-    // placed before stripEmojiWithText because stripEmojiWithText partially removes angle brackets
-    stripTags,
-    lower,
-    trim,
-    stripEmojiWithText,
-    singularize,
+    thoughtTransform,
     ...globals.disableThoughtHashing ? [] : [murmurHash3.x64.hash128],
   ])(value)
 )
