+++ src/util/index.ts
@@ -92,6 +92,7 @@ export { strip } from './strip'
 export { stripPunctuation } from './stripPunctuation'
 export { subsetThoughts } from './subsetThoughts'
 export { sumSubthoughtsLength } from './sumSubthoughtsLength'
+export { thoughtTransform } from './thoughtTransform'
 export { throttleByAnimationFrame } from './throttleByAnimationFrame'
 export { timeDifference } from './timeDifference'
 export { timestamp } from './timestamp'
